revert route nesting
